chore: Session reuse and improved logging
