+++ auto_update.py
@@ -6,7 +6,7 @@ from datetime import datetime
 
 from commons.utils import datetime_to_iso8601_str
 from dojo import get_latest_git_tag
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 from dojo.utils.config import source_dotenv
 
 source_dotenv()
