+++ commons/block_subscriber.py
@@ -8,7 +8,7 @@ import websockets
 
 from commons.objects import ObjectManager
 from dojo.chain import parse_block_headers
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 
 BLOCK_TIME = 12
 WS_OPEN_TIMEOUT = 30
