+++ commons/human_feedback/dojo.py
@@ -9,7 +9,7 @@ import httpx
 from commons.exceptions import CreateTaskFailed
 from commons.utils import loaddotenv
 from dojo import get_dojo_api_base_url
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 from dojo.protocol import (
     CodeAnswer,
     MultimediaAnswer,
