+++ commons/score_storage.py
@@ -5,7 +5,7 @@ import torch
 
 from database.client import connect_db
 from database.prisma.models import Score_Model
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 
 
 class ScoreStorage:
