+++ commons/scoring.py
@@ -5,7 +5,7 @@ import torch
 from torch.nn import functional as F
 
 from commons.utils import _terminal_plot
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 from dojo.protocol import (
     CompletionResponse,
     CriteriaType,
