+++ commons/utils.py
@@ -19,7 +19,7 @@ from Crypto.Hash import keccak
 from tenacity import RetryError, Retrying, stop_after_attempt, wait_exponential_jitter
 
 from commons.objects import ObjectManager
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 
 ROOT_WEIGHT = 0.18
 ROOT_NETUID = 0
