+++ dojo/utils/uids.py
@@ -10,7 +10,7 @@ from commons.exceptions import FatalSubtensorConnectionError
 from commons.objects import ObjectManager
 from commons.utils import aget_effective_stake, get_effective_stake, keccak256_hash
 from dojo.chain import get_async_subtensor
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 
 
 def is_uid_available(metagraph: bt.metagraph, uid: int) -> bool:
