+++ scripts/create_synthetic_task.py
@@ -4,7 +4,7 @@ import json
 from commons.dataset.synthetic import SyntheticAPI
 from commons.human_feedback.dojo import DojoAPI
 from commons.utils import set_expire_time
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 from dojo.protocol import (
     ScoreCriteria,
     TaskSynapseObject,
