+++ simulator/validator.py
@@ -10,7 +10,7 @@ import dojo
 from commons.dataset.synthetic import SyntheticAPI
 from commons.orm import ORM
 from commons.utils import get_epoch_time, get_new_uuid, set_expire_time, ttl_get_block
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 from dojo.protocol import (
     DendriteQueryResponse,
     FeedbackRequest,
