+++ validator_api/validator_logging/core/service.py
@@ -1,6 +1,6 @@
 from typing import Dict, List
 
-from dojo.logging.logging import logging as logger
+from dojo.logging import logging as logger
 
 from .models import LogEntry, LogResponse
